fix(trading): override fees discounts if aggresor is maker and refactor
